requirements.txt
fastapi
flask
openai
numpy
python-dotenv
uvicorn
sqlalchemy
pyodbc
pandas
